fixed: update file location and imports
Closes #1178
