+++ packages/core/src/service/account-management.ts
@@ -5,21 +5,36 @@ import * as ethUtil from 'ethereumjs-util';
 import { getApp } from 'firebase/app';
 import { getAuth } from 'firebase/auth/web-extension';
 
-import { registerStatusKey } from '@onflow/flow-wallet-data-model/cache-data-keys';
+import {
+  userMetadataKey,
+  mainAccountsKey,
+  registerStatusKey,
+} from '@onflow/flow-wallet-data-model/cache-data-keys';
 import { FLOW_BIP44_PATH } from '@onflow/flow-wallet-shared/constant/algo-constants';
 import {
   HTTP_STATUS_CONFLICT,
   HTTP_STATUS_TOO_MANY_REQUESTS,
 } from '@onflow/flow-wallet-shared/constant/domain-constants';
-import { type AccountKeyRequest } from '@onflow/flow-wallet-shared/types/network-types';
-import { type FlowAddress } from '@onflow/flow-wallet-shared/types/wallet-types';
-import { isValidFlowAddress } from '@onflow/flow-wallet-shared/utils/address';
+import {
+  type AccountKeyRequest,
+  type UserInfoResponse,
+} from '@onflow/flow-wallet-shared/types/network-types';
+import {
+  type FlowAddress,
+  type ProfileBackupStatus,
+} from '@onflow/flow-wallet-shared/types/wallet-types';
+import {
+  isValidFlowAddress,
+  isValidEthereumAddress,
+} from '@onflow/flow-wallet-shared/utils/address';
+import { consoleError } from '@onflow/flow-wallet-shared/utils/console-log';
 
 import { preferenceService } from '.';
 import googleDriveService from './googleDrive';
 import keyringService from './keyring';
 import { mixpanelTrack } from './mixpanel';
 import openapiService from './openapi';
+import userInfoService from './user';
 import userWalletService, {
   addPendingAccountCreationTransaction,
   addPlaceholderAccount,
@@ -30,7 +45,7 @@ import {
   pubKeyAccountToAccountKey,
   pubKeySignAlgoToAccountKey,
 } from '../utils/account-key';
-import { setCachedData } from '../utils/data-cache';
+import { getValidData, setCachedData } from '../utils/data-cache';
 import { findAddressWithPK, findAddressWithSeed } from '../utils/modules/findAddressWithPK';
 import { getOrCheckAccountsByPublicKeyTuple } from '../utils/modules/findAddressWithPubKey';
 import {
@@ -41,7 +56,7 @@ import {
 } from '../utils/modules/publicPrivateKey';
 import { generateRandomId } from '../utils/random-id';
 
-export class accountManagement {
+export class AccountManagement {
   async registerNewProfile(username: string, password: string, mnemonic: string): Promise<void> {
     // The account is the public key of the account. It's derived from the mnemonic. We do not support custom curves or passphrases for new accounts
     const accountKey: AccountKeyRequest = await getAccountKey(mnemonic);
@@ -455,6 +470,229 @@ export class accountManagement {
 
     return [_account];
   }
+
+  /**
+   * Get profile backup statuses for password change operation
+   * @param currentPassword - The current password to test decryption
+   * @returns Promise<ProfileBackupStatus[]> - Array of profile backup statuses
+   */
+  async getProfileBackupStatuses(currentPassword: string): Promise<ProfileBackupStatus[]> {
+    try {
+      // Get all backups from Google Drive
+      const backupLists = await googleDriveService.loadBackupAccountLists();
+      // Get all active profiles
+      const userList = userInfoService.getUserList();
+
+      // Get all keyring ids
+      const keyringIds = await keyringService.getAllKeyringIds();
+
+      // Determine active profiles from the keyring ids
+      const activeProfiles = keyringIds.map((id): UserInfoResponse => {
+        const matchingUser = userList.find((user) => user.id === id);
+        if (!matchingUser) {
+          return {
+            username: `unknown_${id.slice(0, 4)}`,
+            id: id,
+            avatar: '',
+            nickname: 'unknown',
+            private: 0,
+            created: '',
+          };
+        }
+        return matchingUser;
+      });
+
+      // Test decryption for each backup
+      const backupStatuses: ProfileBackupStatus[] = await Promise.all(
+        backupLists.map(async (backup) => {
+          const matchingProfile = activeProfiles.find(
+            (profile) => profile.username === backup.username
+          );
+          const isActive = !!matchingProfile;
+          let canDecrypt = false;
+
+          try {
+            // Attempt to decrypt with current password
+            canDecrypt = await googleDriveService.testProfileBackupDecryption(
+              backup.username,
+              currentPassword
+            );
+          } catch (err) {
+            consoleError(`Cannot decrypt backup for ${backup.username}`, err);
+          }
+
+          return {
+            username: backup.username,
+            uid: backup.uid,
+            id: matchingProfile?.id || '',
+            isActive,
+            isBackedUp: true,
+            canDecrypt,
+            isSelected: canDecrypt, // Pre-select those we can decrypt
+          };
+        })
+      );
+
+      // Add active profiles that aren't backed up
+      activeProfiles.forEach((profile) => {
+        if (!backupStatuses.some((status) => status.username === profile.username)) {
+          backupStatuses.push({
+            username: profile.username,
+            uid: null,
+            id: profile.id,
+            isActive: true,
+            isBackedUp: false,
+            canDecrypt: false,
+            isSelected: false,
+          });
+        }
+      });
+
+      return backupStatuses;
+    } catch (err) {
+      consoleError('Failed to get profile backup statuses:', err);
+      throw new Error('Failed to get profile backup statuses');
+    }
+  }
+
+  /**
+   * Change password with selective profile backup updates
+   * @param currentPassword - The current password
+   * @param newPassword - The new password
+   * @param selectedProfiles - List of profile usernames to update backups for
+   * @param ignoreBackupsAtUsersOwnRisk - Whether to ignore backups (for users without Google permission)
+   * @returns Promise<boolean> - Success status
+   */
+  async changePassword(
+    currentPassword: string,
+    newPassword: string,
+    selectedProfiles: string[] = [],
+    ignoreBackupsAtUsersOwnRisk: boolean = false
+  ): Promise<boolean> {
+    try {
+      // If ignoring backups, just change the wallet password
+      if (ignoreBackupsAtUsersOwnRisk) {
+        return await keyringService.changePassword(currentPassword, newPassword);
+      }
+
+      // Handle Google backups if we have Google permission
+      const hasGooglePermission = await googleDriveService.hasGooglePermission();
+
+      if (hasGooglePermission && selectedProfiles.length > 0) {
+        // First update the Google backups
+        await googleDriveService.setNewPassword(currentPassword, newPassword, selectedProfiles);
+
+        // Only change the keyring password if the backup update succeeds
+        const success = await keyringService.changePassword(currentPassword, newPassword);
+
+        if (!success) {
+          throw new Error('Failed to change wallet password after updating backups');
+        }
+
+        // Track successful password change
+        mixpanelTrack.track('password_updated', {
+          address: (await userWalletService.getCurrentAddress()) || '',
+          success: true,
+          profilesUpdated: selectedProfiles.length,
+        });
+
+        return true;
+      } else {
+        // No backups to update, just change the wallet password
+        const success = await keyringService.changePassword(currentPassword, newPassword);
+
+        if (success) {
+          // Track successful password change
+          mixpanelTrack.track('password_updated', {
+            address: (await userWalletService.getCurrentAddress()) || '',
+            success: true,
+            profilesUpdated: 0,
+          });
+        }
+
+        return success;
+      }
+    } catch (err) {
+      consoleError('Error changing password with backups:', err);
+      mixpanelTrack.track('password_update_failed', {
+        address: (await userWalletService.getCurrentAddress()) || '',
+        error: (err as Error).message,
+      });
+      throw err;
+    }
+  }
+
+  /**
+   * Update account metadata (emoji, name, background color) via openapi API
+   */
+  async updateAccountMetadata(address: string, icon: string, name: string, background: string) {
+    const result = await openapiService.updateAccountMetadata(address, icon, name, background);
+
+    // Update the metadata cache after successful update
+    try {
+      const currentPubKey = userWalletService.getCurrentPubkey();
+      const cacheKey = userMetadataKey(currentPubKey);
+
+      // Get existing metadata from cache
+      const existingMetadata = (await getValidData(cacheKey)) || {};
+      const updatedMetadata = {
+        ...existingMetadata,
+        [address]: {
+          background,
+          icon,
+          name,
+        },
+      };
+
+      // Update the cache with new metadata
+      await setCachedData(cacheKey, updatedMetadata, 300_000);
+
+      // Update the specific account in the main accounts cache
+      try {
+        const network = await userWalletService.getNetwork();
+        const accountsCacheKey = mainAccountsKey(network, currentPubKey);
+        const existingMainAccounts = await getValidData(accountsCacheKey);
+
+        if (existingMainAccounts && Array.isArray(existingMainAccounts)) {
+          const updatedMainAccounts = existingMainAccounts.map((account) => {
+            if (account.address === address) {
+              return {
+                ...account,
+                name: name,
+                icon: icon,
+                color: background,
+              };
+            }
+            //Update evmAccount if the address is a valid EVM address
+            if (
+              account.evmAccount &&
+              isValidEthereumAddress(address) &&
+              account.evmAccount.address === address
+            ) {
+              return {
+                ...account,
+                evmAccount: {
+                  ...account.evmAccount,
+                  name: name,
+                  icon: icon,
+                  color: background,
+                },
+              };
+            }
+            return account;
+          });
+
+          await setCachedData(accountsCacheKey, updatedMainAccounts, 60_000);
+        }
+      } catch (updateError) {
+        consoleError('Failed to update main accounts cache:', updateError);
+      }
+    } catch (error) {
+      consoleError('Failed to update metadata cache:', error);
+    }
+
+    return result;
+  }
 }
 
-export default new accountManagement();
+export default new AccountManagement();
